feat: Summarize long lists in tables
The table view for "Rotate Based on a List of IPs (Multi-Records)" and other lists would break when the "Records" or "IP Pool" columns contained a large number of items.

This change introduces a new `summarize_list` helper function that summarizes lists with more than two items by showing the first and last items separated by "...".

This new function is used in the following places to prevent the table from breaking:
- `list_global_rotations`: for "Records" and "IP Pool" columns.
- `list_records_from_config`: for the "IPs" column.
- `list_rotation_groups`: for the "Records" column.